requirements.txt
asgiref==3.5.2; python_version >= "3.8"
backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.8"
django-cors-headers==3.13.0; python_version >= "3.7"
django==4.1.1; python_version >= "3.8"
djangorestframework-simplejwt==5.2.0; python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.6"
gunicorn==20.1.0; python_version >= "3.5"
pillow==9.2.0; python_version >= "3.7"
pyjwt==2.4.0; python_version >= "3.7"
python-magic==0.4.27; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
pytz==2022.2.1; python_version >= "3.7"
sqlparse==0.4.2; python_version >= "3.8"
tzdata==2022.2; sys_platform == "win32" and python_version >= "3.8"
whitenoise==6.2.0; python_version >= "3.7"
